src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
spotipy==2.25.1
python-dotenv==1.0.1
requests==2.31.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
sqlalchemy==2.0.23
ytmusicapi==0.24.1 